+++ packages/gamut/src/Button/ButtonInner.tsx
@@ -1,4 +1,4 @@
-import { fontSize, fontSmoothing, timing } from '@codecademy/gamut-styles';
+import { fontSmoothing, fontWeight, timing } from '@codecademy/gamut-styles';
 import isPropValid from '@emotion/is-prop-valid';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
@@ -8,11 +8,35 @@ export const ButtonInner = styled('span', {
     isPropValid(prop) && !['mode', 'size'].includes(prop),
 })(() => {
   return css`
-    display: flex;
-    font-size: ${fontSize[16]};
+    ${fontSmoothing()}
+    display: inline-block;
+    border: 2px solid transparent;
+    border-radius: 3px;
+    font-weight: ${fontWeight.title};
+    text-align: center;
     transition: ${timing.fast} background-color, ${timing.fast} box-shadow,
       ${timing.fast} color;
 
-    ${fontSmoothing()}
+    &:before,
+    &:after {
+      content: '';
+      display: inline-block;
+      width: 1px;
+      height: 100%;
+      margin-left: -1px;
+      margin-right: -1px;
+      vertical-align: middle;
+    }
+
+    > * {
+      vertical-align: middle;
+
+      &:first-of-type {
+        margin-left: -1px;
+      }
+      &:last-of-type {
+        margin-right: -1px;
+      }
+    }
   `;
 });